requirements.txt
annotated-types==0.7.0
anyio==4.6.0
APScheduler==3.10.4
azure-core==1.31.0
azure-storage-file-share==12.19.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
cryptography==43.0.3
Deprecated==1.2.14
distro==1.9.0
Flask==3.0.3
Flask-Limiter==3.8.0
Flask-Session==0.8.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonalias==0.1.1
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgspec==0.18.6
mysql-connector-python==9.0.0
openai==1.51.0
ordered-set==4.1.0
packaging==24.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
regex==2024.9.11
requests==2.32.3
rich==13.9.1
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
solana==0.34.3
solders==0.21.0
soupsieve==2.6
tiktoken==0.8.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websockets==11.0.3
Werkzeug==3.0.4
wrapt==1.16.0
